# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para a plataforma SkillBridge — liga estudantes e projetos.",
        "title": "SkillBridge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/donations/webhook": {
            "post": {
                "description": "Recebe eventos do Stripe (payment_intent.succeeded / payment_intent.payment_failed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "donations"
                ],
                "summary": "Webhook Stripe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Assinatura do Stripe",
                        "name": "Stripe-Signature",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}